+++ cmd/shortener/output.txt
@@ -26,3 +26,8 @@
 {"short_url":"http://127.0.0.1:8080/26","original_url":"https://github.com/","userID":"9321e158-b3db-4f27-878a-224a756cb387","correlation_id":""}
 {"short_url":"http://127.0.0.1:8080/XVlBz","original_url":"https://github.com/","userID":"9321e158-b3db-4f27-878a-224a756cb387","correlation_id":""}
 {"short_url":"http://127.0.0.1:8080/gbaiC","original_url":"https://github.com/","userID":"9321e158-b3db-4f27-878a-224a756cb387","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/29","original_url":"https://githulssb.com/","userID":"e77dc8ea-4013-4185-8eda-24112519aa8e","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/30","original_url":"https://githulssb.com/","userID":"e77dc8ea-4013-4185-8eda-24112519aa8e","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/31","original_url":"https://githulssb.com/","userID":"e77dc8ea-4013-4185-8eda-24112519aa8e","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/32","original_url":"https://githulssb.com/","userID":"e77dc8ea-4013-4185-8eda-24112519aa8e","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/33","original_url":"https://github.com/","userID":"e77dc8ea-4013-4185-8eda-24112519aa8e","correlation_id":""}
