+++ cmd/shortener/texts.txt
@@ -26,3 +26,8 @@ https://github.com/
 https://github.com/
 https://github.com/
 https://github.com/
+https://githulssb.com/
+https://githulssb.com/
+https://githulssb.com/
+https://githulssb.com/
+https://github.com/
