+++ internal/repository/postgresql.go
@@ -2,14 +2,20 @@ package repository
 
 import (
 	"database/sql"
+	"fmt"
 
+	"github.com/AltynayK/firstpraktikum/internal/models"
 	_ "github.com/lib/pq"
 )
 
 type Config struct {
 	DBdns *string
 }
+type DataBase struct{}
 
+func NewDataBase() Repo {
+	return &DataBase{}
+}
 func NewPostgresDB(cfg Config) (*sql.DB, error) {
 	db, err := sql.Open("postgres", *cfg.DBdns)
 	if err != nil {
@@ -30,4 +36,37 @@ func CreateTable(db *sql.DB) {
 	}
 	DB = db
 
+}
+func (d *DataBase) InsertData(shortURL string, originalURL string, userID string) bool {
+	sqlStatement := `INSERT INTO data (short_url, original_url, user_id) VALUES ($1, $2, $3)`
+	_, err := DB.Exec(sqlStatement, shortURL, originalURL, userID)
+	return err == nil
+}
+
+func (d *DataBase) InsertMultipleData(shortURL string, originalURL string, userID string, correlationID string) bool {
+	sqlStatementt := `INSERT INTO data (short_url, original_url, user_id, correlation_id) VALUES ($1, $2, $3, $4)`
+	_, err := DB.Exec(sqlStatementt, shortURL, originalURL, userID, correlationID)
+	return err == nil
+}
+
+func (d *DataBase) GetLongURLByID(id int) string {
+	row := DB.QueryRow("SELECT original_url FROM data WHERE id = $1", id)
+	alb := models.DBUrl{}
+	if err := row.Scan(&alb.Originalurl); err != nil {
+		fmt.Print("Error.")
+	}
+	return alb.Originalurl
+}
+func ReturnShortURL(LongURL string) string {
+	row := DB.QueryRow("SELECT short_url FROM data WHERE original_url = $1", LongURL)
+	alb := models.DBUrls{}
+	if err := row.Scan(&alb.Shorturl); err != nil {
+		fmt.Print("Error.")
+	}
+	return alb.Shorturl
+}
+
+func Ping() bool {
+	err := DB.Ping()
+	return err == nil
 }
