+++ output.txt
@@ -1,0 +1,2 @@
+{"short_url":"http://127.0.0.1:8080/1","original_url":"https://github.com/","userID":"9321e158-b3db-4f27-878a-224a756cb387","correlation_id":""}
+{"short_url":"http://127.0.0.1:8080/2","original_url":"https://github.com/","userID":"9321e158-b3db-4f27-878a-224a756cb387","correlation_id":""}
